Fix topology port connection/direct topology resolution ordering
