+++ compiler/lib/src/main/scala/analysis/Analysis.scala
@@ -72,7 +72,7 @@ case class Analysis(
   /** The interface under construction */
   interface: Option[Interface] = None,
   /** The map from topology symbols to 'partial' topologies
-   * with only port interface/instance  information */
+   * with only port interface/instance information */
   partialTopologyMap: Map[Symbol.Topology, Topology] = Map(),
   /** The map from topology symbols to topologies */
   topologyMap: Map[Symbol.Topology, Topology] = Map(),
