+++ compiler/lib/src/main/scala/analysis/CheckSemantics/CheckSemantics.scala
@@ -32,6 +32,7 @@ object CheckSemantics {
       a <- CheckComponentInstanceDefs.visitList(a, tul, CheckComponentInstanceDefs.transUnit)
       _ <- CheckComponentInstanceDefs.checkIdRanges(a)
       a <- CheckStateMachineDefs.visitList(a, tul, CheckStateMachineDefs.transUnit)
+      a <- CheckTopologyInstances.visitList(a, tul, CheckTopologyInstances.transUnit)
       a <- CheckTopologyDefs.visitList(a, tul, CheckTopologyDefs.transUnit)
       a <- BuildSpecLocMap.visitList(a, tul, BuildSpecLocMap.transUnit)
       a <- CheckSpecLocs.visitList(a, tul, CheckSpecLocs.transUnit)
